Move get_char_slice and is_valid_encoding to implementation classes
